Include inline theming parameters
A new `option` has been added to specify a library name.
In case it is set a rule will be added to the bottom of the CSS
containing the `variables` as data-uri.
This allows UI5 to evaluate this content and use the JSON without
having to load a separate file via XHR.