Set default of parser option "relativeUrls" to "true"
This is a breaking change!
In order to get the old behavior back, please set the
"parser.relativeUrls" option to "false".

As ui5 themes usually import other themes (e.g. base) to inherit
variables, mixins and rules it is required to set the "relativeUrls"
option to "true" in order to generate valid URLs (e.g. images/fonts).

The default in LESS 1.6.3 is "false".