Reduce collected variables to only add relevant ones
When collecting the variables used during compilation there is a high
chance for redundancy when multiple compiled files (e.g. library themes)
are using the same base files for sharing variables.
With this change those variables will be excluded to reduce the size of
the variables / inline theming parameters.

The full set of variables is available as "allVariables" in the result
object.