[BREAKING] Only rewrite image paths for RTL-variant when files exist (#162)
Only rewrite img path to img-RTL in RTL (right-to-left) CSS if image file is present in img-RTL folder.
Don't rewrite img paths with a protocol (http/https/data/...) or starting with a slash (`/`).

BREAKING CHANGE:

This affects the output of the `rtl` (right-to-left) variant that is created by applying several mechanisms to create a mirrored variant of the CSS to be used in locales that use a right to left text direction.

One aspect is adopting urls which contain an `img` folder in the path.
Before this change, all urls have been changed to use a `img-RTL` folder instead. This allows mirrored images to be used, but it also requires an images to be available on that path even when the original image should be used (e.g. for a logo).

With this change:
- Urls are only adopted in case an `img-RTL` variant of that file exists
- Urls with a protocol (http/https/data/...) or starting with a slash (`/`) are not adopted anymore

Co-authored-by: Matthias Osswald <mat.osswald@sap.com>